find_dominant_color/requirements.txt
numpy==1.21.2
Pillow==8.3.2
scipy==1.7.1
webcolors==1.11.1